# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Building & Industrial IoT API
  description: |
    Unified API for Industrial IoT machine monitoring and Smart Building environmental sensing.
    
    **Data Sources:**
    - Machines module: Based on synthetic industrial IoT dataset (temps, vibration, power, maintenance)
    - Rooms module: Based on KETI Smart Building dataset (CO2, humidity, luminosity, PIR occupancy)
    
    **Base URL:** `http://localhost:3000`
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/your-org/simulated-smartbuilding-api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production
paths:
...
  /api/v1/machines/{machineId}/readings:
    get:
      tags:
        - Sensor Readings
      summary: Get paginated historical sensor readings for a machine
      operationId: getMachineReadings
      parameters:
        - name: machineId
          in: path
          required: true
          description: Machine UUID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: startTime
          in: query
          description: ISO 8601 start timestamp
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          description: ISO 8601 end timestamp
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated readings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorReadingListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/machines/{machineId}/readings/latest:
    get:
      tags:
        - Sensor Readings
      summary: Get most recent persisted sensor reading
      operationId: getLatestMachineReading
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Latest reading
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorReading'
              examples:
                success:
                  value:
                    id: "660e8400-e29b-41d4-a716-446655440001"
                    machineId: "550e8400-e29b-41d4-a716-446655440000"
                    timestamp: "2026-03-23T12:30:45Z"
                    temperatureC: 45.2
                    vibrationMs2: 3.5
                    soundDb: 78.9
                    powerKw: 12.5
                    oilLevelPct: 85.0
                    coolantLevelPct: 90.0
                    laserIntensityW: null
                    hydraulicPressureBar: null
                    spindleSpeedRpm: 2500
                    toolWearMm: 0.15
                    pressureBar: null
                    flowRateLMin: null
                    runtimeHours: 1250.5
                    cycleCount: 5432
                    isAnomaly: false
                    remainingUsefulLifeDays: 245
                    failureWithin7Days: false
                    createdAt: "2026-03-23T12:30:45Z"
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/machines/{machineId}/readings/generate:
    post:
      tags:
        - Sensor Readings
      summary: Generate and save N simulated readings
      operationId: generateMachineReadings
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReadingsRequest'
      responses:
        '201':
          description: Readings generated and saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 100
                  interval:
                    type: number
                    example: 10
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/sensor-readings/anomalies:
    get:
      tags:
        - Sensor Readings
      summary: Get all anomalous readings across all machines
      operationId: getAnomalies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Anomalous readings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorReadingListResponse'

  # ============================================================================
  # MAINTENANCE RECORDS MODULE
  # ============================================================================
...
components:
  schemas:
...
    GenerateReadingsRequest:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 2000
          default: 1
          description: Number of readings to generate
          example: 100
        intervalSec:
          type: number
          minimum: 1
          default: 10
          description: Simulated interval between readings in seconds
          example: 10
...
    SensorReading:
      type: object
      description: Industrial machine sensor reading (temperature, vibration, power, etc.)
      properties:
        id:
          type: string
          format: uuid
        machineId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        temperatureC:
          type: number
          description: Temperature in Celsius
        vibrationMs2:
          type: number
          description: Vibration in mm/s²
        soundDb:
          type: number
          description: Sound level in decibels
        powerKw:
          type: number
          description: Power consumption in kilowatts
        oilLevelPct:
          type: number
          description: Oil level 0–100%
        coolantLevelPct:
          type: number
          nullable: true
          description: Coolant level (null if not applicable)
        laserIntensityW:
          type: number
          nullable: true
          description: Laser intensity (only for LASER_CUTTER)
        hydraulicPressureBar:
          type: number
          nullable: true
          description: Hydraulic pressure (HYDRAULIC_PRESS, INJECTION_MOLD)
        spindleSpeedRpm:
          type: number
          nullable: true
          description: Spindle speed (CNC machines)
        toolWearMm:
          type: number
          nullable: true
          description: Tool wear in millimeters
        pressureBar:
          type: number
          nullable: true
          description: System pressure
        flowRateLMin:
          type: number
          nullable: true
          description: Flow rate in liters per minute
        runtimeHours:
          type: number
          description: Machine runtime at this reading
        cycleCount:
          type: integer
          description: Cumulative cycles
        isAnomaly:
          type: boolean
          description: Detected anomaly flag
        remainingUsefulLifeDays:
          type: number
          nullable: true
        failureWithin7Days:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SensorReadingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SensorReading'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 400
            message: "Validation failed"
            error: "Bad Request"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 404
            message: "Machine not found"
            error: "Not Found"
...